Update breeze.ajaxpost.js
This change was required after upgrading to breeze 1.4.12, otherwise the ajax function call has the "this" keyword undefined inside the breeze library. For my understanding the "this" should be the ajaxAdapter.

Fault code in the breeze library is at line 14952:
  if (!core.isEmpty(this.defaultSettings)) {

Without this change, the "this" keyword is undefined on breeze 1.4.12 release, it was ok on 1.4.11 release.